==> scripts/00-model-funs.Rmd <==
---
title: "Helper Functions for Predicting Temperatures"
author: "Charlotte Mann GSI"
date: "`r Sys.Date()`"
output: github_document
---

```{r setup-mod, include=FALSE}
knitr::opts_chunk$set(echo = FALSE)
```

## Predict from yesterday's temperature for all given future times

```{r}
predict_temp_naive <- function(current.temps, days_lag, date){
  
  preds <- current.temps %>%
    select(station_code, pred.tmin = min_temp,
           pred.tmax = max_temp, pred.tavg = mean_temp) %>%
    mutate(date_pred = date + days(days_lag)) %>%
    select(station_code, date_pred, pred.tmin, pred.tavg, pred.tmax)

  
  return(preds)
  
}

```

## Predict from historical mean statistic for the given day

```{r}

predict_temp_hist_mean <- function(hist.temps, days_lag, date){
  
  pred_date = date + days(days_lag)
  yr = year(date)
  
  preds <- hist.temps %>%
    mutate(month_day = make_date(yr, month(date), day(date))) %>%
    filter(month_day == pred_date) %>%
    group_by(station_code, month_day) %>%
    summarize(across(starts_with("t"), ~mean(.x, na.rm = T), .names = "pred.{.col}")) %>%
    rename(date_pred = month_day) %>%
    select(station_code, date_pred,  pred.tmin, pred.tavg, pred.tmax)
  
  return(preds)
  
}


```



## Predict from historical data - 2 weeks around today's date (local=T) or all days (local = F)


Function to fit a model either on all historical data, or for historical data 2 weeks surrounding date of interest for prediction.

Models are of the form:

$$Y_{s,t,d} = \alpha + \beta Y_{s,t,(d-\ell)} + \alpha_{s}I_s + \beta_{s}Y_{s,t,(d-\ell)}I_s + \epsilon_{s,t,d}$$
Where $s$ is the station id, $t$ is the year, $d$ is the day of the year, and $\ell$ is the number of lag days. $I_s$ is an indicator for the station. This is equivalent to fitting 20 different models - it just makes things easier with the data setup.


```{r}
fit_hist_model <- function(hist.temps, days_lag, date){
      
  startm<-month(date)
  startd<-day(date)
  
  mod.dat <- hist.temps %>%
      mutate(year = year(date)) %>%
      mutate(base_date = make_date(year, startm, startd),
             start_date = base_date - weeks(1) - days(days_lag),
             end_date = start_date + weeks(2) + days(days_lag)) %>%
      mutate(keep = date %within% interval(start_date, end_date))%>%
      filter(keep) %>%
      group_by(station_code, year) %>%
      mutate(across(starts_with("t"), ~lag(.x, n = days_lag), .names = "{.col}_lag"))

  tmin.mod <- lm(tmin ~ tmin_lag*station_code, data= mod.dat)
  tmax.mod <- lm(tmax ~ tmax_lag*station_code, data= mod.dat)
  tavg.mod <- lm(tavg ~ tavg_lag*station_code, data= mod.dat)

  return(list(tmin = tmin.mod, tmax = tmax.mod, tavg = tavg.mod))

}
```

Make predictions based on historical data. When predicting only using the weeks surrounding the date,
fit that model. When using all historical data, give the function those model fits (they don't need to be run over and over)

```{r}
predict_temp_hist <- function(current.temps, date, hist.temps, days_lag){
  
  mods <- fit_hist_model(date=date, hist.temps=hist.temps, days_lag = days_lag)
  tmin.model <- mods[["tmin"]]
  tmax.model <- mods[["tmax"]]
  tavg.model <- mods[["tavg"]]
  
  current.temps <- current.temps %>%
    rename(tmin_lag = min_temp,
           tmax_lag = max_temp,
           tavg_lag = mean_temp)
  
  pred.tmin <- predict(tmin.model, newdata = current.temps)
  pred.tmax <- predict(tmax.model, newdata = current.temps)
  pred.tavg <- predict(tavg.model, newdata = current.temps)
  
  preds <- data.frame(station_code = current.temps$station_code,
                      date_pred = date + days(days_lag),
                      pred.tmin,
                      pred.tavg,
                      pred.tmax)
  
  return(preds)
} 
```

Format predictions for the course specifications.

```{r}
format_preds <- function(preds, date, station_codes_order,
                         fpath){
  
  out <- data.frame(date = as.character(tdy_date))

  for(s in station_codes_order){
    for(d in unique(preds$date_pred)){
      
      temp <- preds %>%
        ungroup() %>%
        filter(station_code == s & date_pred == d) %>%
        select(starts_with("pred")) %>%
        round(digits =2)
      
      out <- cbind(out, temp)
    }
  }
  
  write.table(out, file = fpath, sep = ",",
            row.names = FALSE, col.names=F)
} 
```
