==> scripts/00-wunderground-scrape.Rmd <==
---
title: "Functions to Download and Clean Wunderground Data"
author: "Charlotte Mann GSI"
date: "`r Sys.Date()`"
output: md_document
---

```{r setup2, include=FALSE}
knitr::opts_chunk$set(echo = FALSE)
```

Adapted from script by Yash Patel found [here](https://github.com/yashpatel5400/stats604-project4/blob/main/raw_data/wunderground_download.py)


Get raw data from wunderground
```{r}
fetch_wunderground <- function(stationid = "KPDX", start_date = "20221001", end_date = NULL){
  
  if(is.null(end_date)){end_date = start_date}
  
  h <- new_handle()
  handle_setheaders(h, "accept"= "application/json, text/plain, */*",
                    "sec-ch-ua"="\"Google Chrome\";v=\"107\", \"Chromium\";v=\"107\", \"Not=A?Brand\";v=\"24\"",
                    "sec-ch-ua-mobile"= "?0",
                    "sec-ch-ua-platform"= "\"macOS\"")
  
  url <- paste0("https://api.weather.com/v1/location/",
                stationid,
                ":9:US/observations/historical.json?apiKey=e1f10a1e78da46f5b10a1e78da96f525&units=e&startDate=",
                start_date,
                "&endDate=",
                end_date
  )
  
  json.dat <- curl_fetch_memory(url, handle = h) %$%
    content %>%
    rawToChar %>%
    prettify() %>%
    fromJSON()
  
  return(json.dat$observations)
}
```

Summarize raw data to get min, mean, and max temps for each date included

```{r}
summarize_wunderground <- function(raw.table){
  
  stopifnot("temp" %in% colnames(raw.table),
            "valid_time_gmt" %in% colnames(raw.table),
            "obs_id" %in% colnames(raw.table) )
  
#wunderground is not consistent with exactly what times are included in a day, so it is safest to just pull one date at a time
  
  dat <- raw.table %>%
    select(station_code = obs_id, valid_time_gmt, temp) %>%
    mutate(date_time = as_datetime(valid_time_gmt),
           day = date(date_time),
           tod = hour(date_time)) %>%
    mutate(date_group = min(day)) %>%
    group_by(station_code, date_group) %>%# this is a hacky fix for now - ONLY USE WITH ONE DATE
    summarize(min_temp = min(temp), max_temp = max(temp), mean_temp = mean(temp),
              first_obs_time = min(date_time), last_obs_time = max(date_time)) %>%
    rename(date = date_group)
  
  return(dat)
  
}

```

Testing

```{r}
weather.dat <- fetch_wunderground(start_date = "20221106", end_date = "20221107")

weather.dat %>%
  knitr::kable(format = "markdown")

summarize_wunderground(weather.dat)
```
